refactor(core): 1 validator per file = no named exports = smaller bundle
