+++ packages/core/lib/validators/int-range.js
@@ -1,0 +1,4 @@
+module.exports = (min = 1, max = Infinity) => value =>
+  typeof value === 'number' &&
+  parseInt('' + value, 10) === value &&
+  value >= min && value <= max
