+++ packages/core/lib/validators/list-of-functions.js
@@ -1,0 +1,3 @@
+const { filter, isArray } = require('../es-utils')
+
+module.exports = value => typeof value === 'function' || (isArray(value) && filter(value, f => typeof f === 'function').length === value.length)
