+++ packages/core/lib/validators/string-with-length.js
@@ -1,0 +1,1 @@
+module.exports = value => typeof value === 'string' && !!value.length
